fix(타이머): TINYINT 타입 INTEGER에서 BYTE 로 형 변환하여 DB와 매핑 (#206)

+++ src/main/java/com/project200/undabang/timer/custom/entity/CustomTimerStep.java
@@ -32,7 +32,7 @@ public class CustomTimerStep {
     @Comment("스텝 순서")
     @NotNull
     @Column(name = "custom_timer_steps_order", nullable = false)
-    private Integer customTimerStepOrder;
+    private Byte customTimerStepOrder;
 
     @Comment("스텝 시간")
     @NotNull
